+++ src/index.ts
@@ -38,36 +38,7 @@ export {
 	useUpdateProjectSettings,
 } from './hooks/projects.js'
 export {
-	getClientQueryKey,
-	getDeviceInfoQueryKey,
-	getIsArchiveDeviceQueryKey,
-} from './lib/react-query/client.js'
-export {
-	getDocumentByDocIdQueryKey,
-	getDocumentByVersionIdQueryKey,
-	getDocumentsQueryKey,
-	getManyDocumentsQueryKey,
 	type WriteableDocument,
 	type WriteableDocumentType,
 	type WriteableValue,
-} from './lib/react-query/documents.js'
-export {
-	getInvitesQueryKey,
-	getPendingInvitesQueryKey,
-} from './lib/react-query/invites.js'
-export {
-	getMapsQueryKey,
-	getStyleJsonUrlQueryKey,
-} from './lib/react-query/maps.js'
-export {
-	getAttachmentUrlQueryKey,
-	getDocumentCreatedByQueryKey,
-	getIconUrlQueryKey,
-	getMemberByIdQueryKey,
-	getMembersQueryKey,
-	getProjectByIdQueryKey,
-	getProjectRoleQueryKey,
-	getProjectSettingsQueryKey,
-	getProjectsQueryKey,
-} from './lib/react-query/projects.js'
-export { ROOT_QUERY_KEY } from './lib/react-query/shared.js'
+} from './lib/types.js'
